股票卖出、bootstrap页面设计
股票卖出功能实现，bootstrap网页设计已使用

+++ src/main/java/com/example/demo/User.java
@@ -8,7 +8,7 @@ public class User {
         this.username=username;
     }
 
-    public void setBalance(int balance){
+    public void setBalance(float balance){
         this.balance=balance;
     }
